CA-107115: Xapi_period_scheduler: fall back to Thread.delay if Delay.wait fails
This ensures that the periodic scheduler does not spin and fill up the logs when
Delay.wait fails for some reason. A reason for failing is due to xapi hitting its
limit of open file descriptors.

If Delay.wait fails, we fall back to Thread.delay, which is less likely to fail.
If even this fails, we break out of the scheduler loop, logging an alarming error
message.

The difference between Delay.wait and Thread.delay is that Delay.wait can be
interrupted by a signal, and Thread.delay cannot. Such a signal is used when adding
a new item to the scheduler queue that expires before the current one.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>