CA-107702: Ensure that networkd explicitly configures all physical interfaces with IP in dom0
This is done by marking the interfaces as "persistent" in networkd. Previously, only bridge
interfaces (xenbr*, xapi*) with IP configuration were made persistent. However, their associated
physical interfaces (eth0 etc.) should receive the same treatment, so that attributes such as the
MTU are configured on these as well.

When a host starts, interfaces with IP config need to be configured before xapi can start,
especially on pool slaves, or if HA is enabled.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>