Add xenstore_watch to the build
xenstore_watch should be functionally identical, but has
been rewritten to work with the new xenstore and xenctrl
libraries.

Note the use of the asynchronous set_watch_callback rather
than the synchronous Xs.read_watchevents, which means we
don't need to explicitly create a thread in the
create_watcher_thread function.

Signed-off-by: John Else <john.else@citrix.com>