CA-104674: remove superfluous genid from VMs
On Xapi startup, in dbsync_master, we now remove genids from all VMs that don't
need them. Currently, only Windows 8 and 2012 require genids. We can tell that
a VM is Win8/12 because it has a vga=std platform flag, so we filter out all of
these VMs before removing genids. Since vga=std is unsupported for all other
VMs, we don't mind leaving their genid as-is.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>