CA-109476: Read rrd plugin timestamp as int64
It's written as an int64 by the plugin library. Also, a standard int
isn't big enough to hold a typical Unix epoch.

This value isn't actually used, rrdd just relies on it changing so
that the plugin payload's checksum changes even if the data source
values stay the same. However, it would be nice if it's correct in
case we do ever want to use it!

Signed-off-by: John Else <john.else@citrix.com>