Add scripts to help maintain version tag and changelog
scripts/tag-version.sh will tag the current commit with the version number
found in the VERSION file. You should have already updated the VERSION file and
committed the changes.

scripts/gen-changelog.sh will prepend the CHANGELOG file with a summary of the
pull request commit message bodies, up to the last tagged commit. You should
edit the resulting CHANGELOG file before committing it, and you should have
already bumped the VERSION file to the next version. This change can be
committed in the same commit as the VERSION bump.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>