CA-108916: Revert to free when apply_edition fails
Previously, when applying a new edition failed (for instance, if an incorrect
license server address is used), xapi will continue to operate under the old
edition's license parameters, even though v6d had released the license back to
the license server. This was a design decsion made on-purpose in order to give
the user the benefit of the doubt.

Since it is now less dangerous for a host to lose a license due to a fat-finger
mistake like this, we'd like to remove this functionality and instead revert to
the free edition when V6client.apply_edition returns an exception. We now
reapply the free edition, and then rethrow the original exception.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>