CA-109300: nm: correctly set persistence on parent bridge of VLAN
A VLAN bridge is never "persistent" (i.e. it is never brought up by
networkd on startup), while the parent bridge of the VLAN may be
persistent (in case it has a management IP address).

The bug was that, when creating a VLAN, it set the persistence of
the parent bridge equal to the persitence of the VLAN.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>