Move Xapi_pci over to use Pci_db and Xapi_pci_helpers
This replaces the need for the pci-info python script and gives a richer
data structure to query. This is going to be useful for the vGPU work.

Also moved out test code from the initialisation of the Xapi_pci_helpers and
Pci_db modules into tests so that they don't barf on stdout on initialisation.
These tests are currently disabled (see skip boolean in each module) so that
they don't barf on build.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>