CP-5811: Add VGPU_type.{max_heads, pGPU_footprint}
Added the above fields to the datamodel. The max_heads is the number of
displays the VGPU type can support and is taken directly from the config file.

The pGPU_footprint is an abstract utilisation tracking metric and is internal
only. It is calculated using the max_instance field from the VGPU config.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>