Do not try deleting the old xenguest log file.
The only time where this will succeed is when performing localhost
migrations.  This action is a deliberate choice on behalf of the user,
suggesting that they are also likely to want the log files.

Specifically, in the case of a localhost migrate, the logs for the transmit
half of the migrate are deleted.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>