CA-82314: networkd: introduce Linklocal6 IPv6 address mode
This commit redefines the IPv6 address mode "None6" to mean "no IPv6 address at
all, not even a link local one", and introduces the new mode Linklocal6, which
means "only IPv6 address, except a link local one".

Previously, it was not possible to get networkd to remove the link-local IPv6
address from an interface.

When a link local IPv6 address is requested, it is derived from the interface's
MAC address as commonly done.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
Imported-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>