xenopsd: read list and the default setting from the config file
xapi startup doesn't fully complete unless it can contact all
the xenopsd instances it wants to manage. It's easier to setup
a working system if the list can be tweaked via a config file
rather than a recompilation.

Although it's possible to override a xenopsd per-VM, it's also
convenient to set a host-wide default via the config file.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>