xenlight: use Xenlight defaults wherever we can
When a field is added to a type of the libxl IDL, the Xenlight OCaml bindings
turn this into an additional field in a record type.  This would generally mean
that we would have to change our code to take the new field into account.

However, the bindings provide "default" functions that return records with
libxl's defaults filled into, and all fields present. We can therefore make our
code work without change by getting default records from Xenlight, and using
the "with" syntax to fill in just the fields we need and know.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>