More fixes for port locking.
The script 'common.py' reads the config from xenopsd's database
directly. The past few commits made sure the on-disk database
was up-to-date after API calls to mutate various fields. However,
there is also some further modification of some of the fields
done by the xc driver - for example, the 'extra_private_fields'
string -> string map is augmented with some extra kvs derived
from other fields. In particular, the ipv4_allowed and
ipv6_allowed keys are derived from the locking_mode parameter.
These keys ended up in xenstore but not in the on-disk
json representation of the VIF. The same happened for the
network-backend field - added by xenopsd to the same map, and
thus not in the on-disk json. This was fixed in a xen-api
commit by explicitly adding it in xapi, which was a bit of
a hack.

Signed-off-by: Jon Ludlam <jonathan.ludlam@eu.citrix.com>