VM migration: block the HTTP handler until the xenops migration task has finished
It should have been possible to return from the handler after the
fd had been transferred to xenopsd, but triggered unexplained bad
behaviour. Perhaps the webserver is corrupting the memory stream?

With this workaround, migration seems to work with the caveat that
a file descriptor is leaked every time a VM is received on a host
(when the HTTP handler wraps the file descriptor as a channel and
transfers it, it seems to leave something behind.)

Signed-off-by: David Scott <dave.scott@eu.citrix.com>