CP-6257: Update GPU_group.[enabled|supported]_VGPU_types at startup
If, when xapi starts up, the Dbsync code finds that new GPUs have
arrived or old GPUs have gone away, we need to update the supported and
enabled VGPU types of the groups containing those GPUs.

The update requests are sent to the master as an API call, for
efficiency.

Signed-off-by: John Else <john.else@citrix.com>