CA-116819: xenlight: don't attach storage during VM_restore
When migration a VM, its VBDs need to be attached after VM_restore, during the
VM_restore_devices operation.  This is the case in classic xenopsd, and must
also be the case in the xenlight version, because the sender host in a
migration detaches itself from the storage only after sending the memory image
across.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>