CA-112882: Allow pool-join without v6d licensing daemon.
When xapi starts, if it cannot contact the licensing daemon it now
falls back to an edition-string of "free/libre" rather than "free".

A host is now allowed to join a pool if the edition-strings of all
hosts in the pool match the edition-string of the host that wants to
join.  Only if there is some mismatch do we consult v6d to work out
which different edition-strings are compatible.

Thus a pool of hosts with no licensing daemons can now be created.

This change also introduces a new error type for license-edition
mismatch between host and pool.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>