CA-118195: xenlight: set O_CLOEXEC when opening a file for libxl
This fixes VM resume for HVM guests: it seems that when libxl forks qemu, qemu
inherits the suspend image fd that was opened by xenopsd, and keeps it open.
This makes it impossible for xenopsd to unmount and cleanup the mount that was
created for the suspend image, after the resume completes.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>