SCTX-1585: Handle hosts having different lists of features
During an upgrade, the older hosts will not indicate whether or not they
support newer features. If a host doesn't support a feature, always
treat as restricted on that host for the purposes of working out the
pool restrictions.

Signed-off-by: John Else <john.else@citrix.com>
(cherry picked from commit fb82876687e7e6d30f7707308914866a61fe9482)