CA-115521: Change the way we use the platform map for vGPU setup
Instead of setting platform:vga=vgpu, xapi now just sets
platform:vgpu_pci_id and platform:vgpu_config. The presence of both of
these keys indicates that the VM's display mode needs to be Vgpu, and
the value of platform:vga is overridden.

This means that when a vGPU is removed from a VM, the display emulation
mode will revert back to what it was before the vGPU was added.

Signed-off-by: John Else <john.else@citrix.com>