CP-6853: Add a means of passing arbitrary arguments to vgpu
The contents of VM.platform:vgpu_extra_args (if any) will be passed to
the vgpu process (and in turn to the NVIDIA plugin).

This change also removes the vgpu_vnc_enabled key, since it can be
passed via the new mechanism (and is also mostly obsolete, since VNC
disabling is now automatic).

Signed-off-by: John Else <john.else@citrix.com>