VDI.copy: allow copying only .vhd deltas
Allow the user to manually specify an identical pair of disks
(one on the source SR and one on the destination SR) so that the
copy can transfer only .vhd deltas.

Usage:

$ xe vdi-copy uuid=<disk-to-copy> sr-uuid=<destination-sr>  \
    driver-params:base-src=<base-disk-on-source-SR>         \
    driver-params:base-dst=<base-disk-on-destination-SR>

where <base-disk-on-source-SR> and <base-disk-on-destination-SR>
must have identical contents (or else the resulting VDI will be
corrupt).

The implementation will clone a VDI from <base-disk-on-destination-SR>
and write .vhd deltas to the new .vhd leaf (via the normal
tapdisk datapath).

Signed-off-by: David Scott <dave.scott@eu.citrix.com>