When attaching a network, only configure managed PIFs
This change means that PIFs that are not managed by xapi can still be used by
VMs, and as management interface, but xapi will assume that they have already
been configured. This means that the interface must be up, it must have a
bridge (with a name matching the related network.bridge field), and it must
have IP configuration (in case of the management interface).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>