Don't check config file's hash before syncing
The config file hash check is a left-over relic from when we had a flash-memory
based "OEM edition". We would check the hash of the "config files" (aka
/etc/passwd) before writing the modified config files to disk. This was meant
to save wear and tear on flash media.

We no longer make a flash-based OEM edition. Furthermore, we're rewriting the
config_file_sync mechanism to only transmit a single hashed password. We might
as well send this over without hashing, and just write the new password if it's
different than the old one.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>