CA-123007: Hold a mutex inside Device.PCI.unbind_from_nvidia
If multiple threads are trying to unbind different devices in parallel,
there is a chance that some of the nodes under /proc/driver/nvidia/ will
disappear after a thread has retrieved the list of nodes but before that
thread has accessed their contents.

Holding a mutex for the duration of the (find device,lock device,unbind)
sequence means this should no longer happen.

Signed-off-by: John Else <john.else@citrix.com>