CP-6232: Write suspend img signature in xenopsd-xl
So that we can migrate from legacy xenopsd on a classic backend to xenopsd on
a xenlight backend, we need to be able to understand the save signature
("XenSavedDomain\n") that is at the start of the disk put there by the legacy
toolstack.

Since we would have to accomodate this and peeking is impracticle on pipes then
we will adopt this signature in the xenopsd-on-xenlight toolstack.

Note this is distinct from the Qemu save signature which has been discarded in
this work (see ef3bf4b and 9ccef3e).

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>