CA-121019: Loosen restrictions on VGPU_type import
When importing a VGPU_type which doesn't exist locally, import the
record instead of failing the import. We'll end up with a VGPU which has
this type, but no PGPUs which support this type, so the VM with the VGPU
will be unable to start.

The VM will be able to start once a PGPU which supports the type is
available.

Signed-off-by: John Else <john.else@citrix.com>