Set pool role to Master during unit tests
Make sure pool role is correctly set during startup. Since Context's initial
database is created at module init time, we need to do this even before the
unit test harness can positively set the pool role. We check that this is a
unit test context by checking that the executable's name is not 'xapi'. If it's
not, we assume unit test.

This commit also modifies config_misc to return a default string for the Xen
version if we're running in unit test mode and get an error that we can't
connect to the xc interface.

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>