Add a 'taskOwner' who must recognise a task to keep it alive
The creator of a task is responsible for destroying it. To highlight
bugs where a task creator fails in their duty, we add the concept of
a task 'owner'. The owner must acknowledge their interest in active
tasks every 30s or so, otherwise the tasks are 'destroy' (i.e.
set to automatically disappear on completion). Note that 'destroy'
only removes the object *when the activity has finished*. It does
not auto-cancel.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>