CA-120846: fix get_mac for bonded interfaces
When an interface is bonded by the Linux bonding driver, the driver may
change the MAC address of the interface to that of the bond. The only
place where you can find the "real" MAC address of a bond slave seems to
be /proc/net/bonding.

Imported from xcp-networkd.git e976b03bbf5262cfb3c0e18382700396dcc54ef6

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
Imported-by: John Else <john.else@citrix.com>