SCTX-1616: Ensure PBD.device_config:SRmaster never exists
This should never be part of the PBD record. Instead it is computed before each
SM call in (the internal function) Sm.__get_my_devconf_for_sr before being
passed to the SM driver. Having this field in the device config will mean that
the wrong value could be passed to the SM.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>